[Feat]/#34 - 지도와 컴포넌트 분리
